+++ Backend/BanQuanAo/src/main/java/com/example/banquanao/repository/CategoryRepository.java
@@ -3,8 +3,11 @@ package com.example.banquanao.repository;
 import com.example.banquanao.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+
 import java.util.List;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     List<Category> findByStatus(String status);
+
+
 }
