+++ Backend/BanQuanAo/src/main/java/com/example/banquanao/service/CategoryService.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 @Service
 public class CategoryService {
+
     private final CategoryRepository repo;
 
     public CategoryService(CategoryRepository repo) {
@@ -17,21 +18,24 @@ public class CategoryService {
     public List<Category> findAll() {
         return repo.findAll(Sort.by(Sort.Direction.ASC, "id"));
     }
-    public Category findById(int id) {
+
+    public Category findById(Long id) {
         return repo.findById(id).orElseThrow();
     }
+
+
     public Category save(Category category) {
         return repo.save(category);
     }
 
-    public Category update(Integer id,Category newCategory) {
+    public Category update(Long id,Category newCategory) {
     Category category = repo.findById(id).orElseThrow();
     category.setName(newCategory.getName());
     category.setDescription(newCategory.getDescription());
     category.setStatus(newCategory.getStatus());
     return repo.save(category);
     }
-    public void delete(Integer id) {
+    public void delete(Long id) {
         Category category = repo.findById(id).orElseThrow();
         category.setStatus("INACTIVE");
         repo.save(category);
